requirements.txt
uvicorn
fastapi
mem0ai
sentence-transformers